+++ hdi/dimensionality_reduction/wtsne.h
@@ -75,7 +75,7 @@ namespace hdi{
       void getEmbeddingPosition(scalar_vector_type& embedding_position, data_handle_type handle)const;
 
       //! Get the number of data points
-      unsigned int getNumberOfDataPoints(){  return _P.size();  }
+      size_t getNumberOfDataPoints(){  return _P.size();  }
       //! Get P
       const sparse_scalar_matrix& getDistributionP()const{ return _P; }
       //! Get Q
